requirements.txt
streamlit==1.22.0
streamlit-extras==0.2.6
numpy==1.24.6
opencv-python==4.8.1.78
Pillow==10.1.0
rembg==2.1.3
onnxruntime==1.15.1